- Fatal error when deleting plugin #889

+++ uninstall.php
@@ -6,10 +6,12 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 }
 
 // Delete options
+if (defined('STAGS_OPTIONS_NAME')) {
 delete_option( STAGS_OPTIONS_NAME ); // Options plugin
 delete_option( STAGS_OPTIONS_NAME . '-version' ); // Version ST
 
 delete_option( STAGS_OPTIONS_NAME_AUTO ); // Options auto tags
+}
 delete_option( 'tmp_auto_tags_st' ); // Autotags Temp
 
 delete_option( 'stp_options' ); // Old options from Simple Tagging !
